+++ ui/ui-expenses/src/store/drivers/index.js
@@ -1,8 +1,10 @@
 /* eslint-env browser */
 import router from '@k-ramel/driver-redux-little-router'
 import routes from './routes'
+import pica from './pica'
 
 export default {
   router: router(routes),
   window: { getDriver: () => window },
+  pica,
 }
